refactor jbrowse
