+++ src/components/Layout/LayoutView.js
@@ -22,8 +22,9 @@ import StudyDetail from '../../pages/study/studyDetailController';
 import UnifiedView from '../../pages/dashboardTab/unifiedViewController';
 import OverlayWindow from '../OverlayWindow/OverlayWindow';
 import GraphqlClient from '../GraphqlClient/GraphqlView';
-import JBrowseDetail from '../../pages/JbrowseDetail/JbrowseController';
+import JBrowseDetail from '../../pages/JbrowseDetail/SingleView/JbrowseController';
 import ModelExplorer from './utils';
+import JbrowseMultiViewController from '../../pages/JbrowseDetail/MultiView/JbrowseController';
 
 const ScrollToTop = () => {
   window.scrollTo(0, 0);
@@ -60,6 +61,7 @@ const Layout = ({ classes, isSidebarOpened }) => (
             <Route path="/program/:id" component={ProgramDetail} />
             <Route path="/case/:id" component={CaseDetails} />
             <Route path="/fileViewer/:type/:id" component={JBrowseDetail} />
+            <Route path="/multiFileViewer" component={JbrowseMultiViewController} />
             <Route path="/icdc-data-model" component={ModelExplorer} />
             {aboutPageRoutes.map(
               (aboutPageRoute) => <Route path={aboutPageRoute} component={About} />,
